Update bots.sql
Dropped support for Byteball mostly due to its lack of stability but also for the lack of multi-platform support (crashes on Arch linux). The persistently low popularity of Byteball gives no reason to keep spending time and effort on coping with the ever-changing witness list and mandatory updates to the base protocol.